requirements.txt
spotipy==2.23.0
python-dotenv==1.0.1
rich==13.7.0